python/projects/my-email/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "my-email"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-auth>=2.28",
    "google-auth-oauthlib>=1.2",
    "google-api-python-client>=2.120",
    "beautifulsoup4>=4.12",
    "html2text>=2024.2",
    "openai>=1.30",
    "click>=8.1",
    "python-dotenv>=1.0",
    "structlog>=24.1",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-mock>=3.14",
    "ruff>=0.4",
]

[project.scripts]
my-email = "my_email.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
